backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.7.0
firebase-admin>=4.5,<6.0
drf-firebase-auth==1.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow>=10.0.0
